[node-api] make metadata a single no_std compatible crate
